Restructure CurrentUserContext
Login/logout prompting correct display of current user on Home
works!

Three possible explanations: the restructure of the currentUser from a state
to a context element, and/or moving the logout currentUser logic
to occur before the fetch, and/or changin the conditions
on the useEffect fetch inside Home.
