+++ app/javascript/contexts/CurrentUserContext.js
@@ -1,7 +1,12 @@
-import React from "react";
+import React, { createContext, useState } from "react";
 
-const CurrentUserContext = React.createContext({})
+export const CurrentUserContext = createContext();
 
-export const CurrentUserProvider = CurrentUserContext.Provider;
-
-export default CurrentUserContext;
+export const CurrentUserProvider = ({ children }) => {
+  const [currentUser, setCurrentUser] = useState(null);
+  return (
+    <CurrentUserContext.Provider value={{currentUser, setCurrentUser}}>
+      {children}
+    </CurrentUserContext.Provider>
+  )
+}
